Add files via upload
this file is a test of the rfind method in python. saw it in the background of a video about data structures and got curious. some extra stuff for finding the next "." and for reversing the string. used os module for clearing the terminal to keep the UX of this neat and readable.